refactor: organize frontend data
