--- frontend/helpers/statesmen.ts
@@ -1,13 +1,0 @@
-export const SERIES_NULLIFIERS: Record<string, string> = {
-  "1a": "Punic",
-  "18a": "Macedonian",
-  "19a": "Macedonian",
-}
-
-export const STATESMAN_ABILITIES: Record<string, string> = {
-  "1a": "Punic War specialist",
-  "2a": "Halves all combat losses",
-  "18a": "Macedonian War specialist",
-  "19a": "Macedonian War specialist",
-  "22a": "1 free tribune per turn",
-}
